FEAT: Updated Code to show Color @ Specific Spot
The Color Identifier now handles color identification
of clicking in specific spots of an image, rather than
simply taking all of the colors of the image and determining
the prominent color. This allows for more accurate identification
as well as making the tool more useful for identifying exact
color codes as provided at specific areas of the iamge.
While more fine-tuning of click area may be required,
this version works to identify a color code at the clicked spot.